feat(data-store-namespace): moved data store classes to Context namespace #73
