+++ Attest.Testing.SpecFlow/ScenarioContextKeyValueDataStoreAdapter.cs
@@ -1,4 +1,4 @@
-﻿using Attest.Testing.DataStore;
+﻿using Attest.Testing.Context;
 using TechTalk.SpecFlow;
 
 namespace Attest.Testing.SpecFlow
