+++ Attest.Testing.SpecFlow/UseKeyValueDataStoreMiddleware.cs
@@ -1,4 +1,4 @@
-﻿using Attest.Testing.DataStore;
+﻿using Attest.Testing.Context;
 using Solid.Bootstrapping;
 using Solid.Practices.Middleware;
