+++ Attest.Testing.xUnit/EndToEndTestsBase.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-using Attest.Testing.DataStore;
+using Attest.Testing.Context;
 
 namespace Attest.Testing.xUnit
 {
