+++ Attest.Testing.xUnit/IntegrationTestsBase.cs
@@ -1,6 +1,6 @@
 ﻿using System;
+using Attest.Testing.Context;
 using Attest.Testing.Core;
-using Attest.Testing.DataStore;
 using Attest.Testing.Integration;
 using Solid.Bootstrapping;
 using Solid.Core;
